% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{add_xlsx_sheet}
\alias{add_xlsx_sheet}
\title{Add sheet to existing Excel workbook}
\usage{
add_xlsx_sheet(wb_location, newsheet_name, newsheet_data)
}
\arguments{
\item{wb_location}{Character string with path to Excel workbook}

\item{newsheet_name}{Character string with name for new sheet}

\item{newsheet_data}{Data frame to write to new sheet}
}
\value{
NULL (modifies workbook file in place)
}
\description{
Adds a new sheet to an existing Excel workbook. The sheet must not exist previously.
}
