% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BNF}
\alias{BNF}
\title{Biological Nitrogen Fixation Parameters}
\format{
A data frame with BNF coefficients:
\describe{
  \item{Name_BNF}{Crop or system name}
  \item{Ndfa}{Proportion of nitrogen derived from atmosphere (0-1)}
  \item{Leguminous_share}{Fraction of leguminous species (0-1)}
  \item{BGN}{Below-ground nitrogen parameter}
  \item{NHI}{Nitrogen harvest index}
}
}
\source{
Scientific literature compilation
}
\description{
Parameters for calculating biological nitrogen fixation by crop type,
including symbiotic and non-symbiotic fixation.
}
\keyword{datasets}
