% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npp_functions.R
\name{Calc_CropNPP_components}
\alias{Calc_CropNPP_components}
\title{Calculate Cropland NPP Components Including Weeds}
\usage{
Calc_CropNPP_components(Crop_NPPpot)
}
\arguments{
\item{Crop_NPPpot}{A data frame with potential NPP values and crop data}
}
\value{
A data frame with complete cropland NPP components (crop + weeds) in DM, C, N
}
\description{
Calculates complete cropland NPP including weed biomass, scaled by potential NPP
and considering fallow periods.
}
\examples{
\dontrun{
cropland_npp <- Calc_CropNPP_components(crop_npp_potential)
}
}
