% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{fill_na_with_sum}
\alias{fill_na_with_sum}
\title{Fill gaps with accumulated sum}
\usage{
fill_na_with_sum(col1, col2)
}
\arguments{
\item{col1}{Column to fill (will be modified)}

\item{col2}{Column with values to add}
}
\value{
Vector with filled values
}
\description{
Fills gaps in a column by accumulating values from another column.
The values are accumulated along the series.
}
\details{
For each NA in col1, fills with col1[i-1] + col2[i]
}
