% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Exclude operator}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{Vector or NULL: the values to be matched}

\item{y}{Vector or NULL: the values to be matched against}
}
\value{
A logical vector indicating if there is no match for each element of x
}
\description{
Negation of the `%in%` operator
}
\examples{
c(1, 2, 3) \%!in\% c(2, 4, 6) # Returns TRUE FALSE TRUE

}
