% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GWP}
\alias{GWP}
\title{Global Warming Potentials}
\format{
A data frame with GWP values:
\describe{
  \item{Gas}{Greenhouse gas name}
  \item{GWP_20}{20-year global warming potential}
  \item{GWP_100}{100-year global warming potential (most commonly used)}
  \item{GWP_500}{500-year global warming potential}
}
}
\source{
IPCC Assessment Reports
}
\description{
GWP values for different greenhouse gases and time horizons based on IPCC guidelines.
}
\keyword{datasets}
