% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IPCC_residue_coefs}
\alias{IPCC_residue_coefs}
\title{IPCC Crop Residue Coefficients}
\format{
A data frame with columns:
\describe{
  \item{IPCC_crop}{IPCC crop category name.}
  \item{Slope_AG}{Slope of the linear residue model (Mg DM residue per
    Mg DM yield per hectare).}
  \item{Intercept_AG_MgDMha}{Intercept (Mg DM / ha), representing
    structural baseline residue production.}
  \item{RS_ratio_IPCC}{Default root-to-shoot ratio from IPCC.}
  \item{N_AG_residue}{Nitrogen content of above-ground residue
    (kg N / kg DM).}
  \item{N_BG_residue}{Nitrogen content of below-ground residue
    (kg N / kg DM).}
  \item{Source}{Literature source reference.}
}
}
\source{
IPCC (2019) 2019 Refinement to the 2006 IPCC Guidelines,
  Vol.4, Ch.11, Table 11.1a. Bolinder et al. (2007). Gan et al. (2009).
}
\description{
Linear model coefficients for estimating above-ground crop residue biomass
from yield, based on IPCC 2019 Refinement guidelines (Vol.4, Ch.11,
Table 11.1a).
}
\keyword{datasets}
