% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IPCC_root_coefs}
\alias{IPCC_root_coefs}
\title{IPCC Root Biomass Coefficients}
\format{
A data frame with columns:
\describe{
  \item{IPCC_crop}{IPCC crop category name.}
  \item{RS_default}{Default root-to-shoot ratio.}
  \item{RS_low_N}{RS ratio under low N input (<60 kg N/ha).}
  \item{RS_high_N}{RS ratio under high N input (>120 kg N/ha).}
  \item{RS_irrigated}{RS ratio under irrigation.}
  \item{RS_rainfed}{RS ratio under rainfed conditions.}
  \item{BG_ref_MgDMha}{Reference below-ground biomass (Mg DM / ha).}
  \item{Source}{Literature source reference.}
}
}
\source{
IPCC (2006) Guidelines for National Greenhouse Gas Inventories,
  Vol.4, Ch.11. Bolinder et al. (2007). Poorter & Nagel (2000).
  Benjamin et al. (2014).
}
\description{
Root-to-shoot ratios and reference root biomass values for estimating
below-ground crop biomass, with adjustments for N input and irrigation.
}
\keyword{datasets}
