% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{items_full}
\alias{items_full}
\title{Item Classifications}
\format{
A data frame with item information:
\describe{
  \item{item_code}{Standardized item code}
  \item{item_name}{Item name}
  \item{category}{Product category}
  \item{group}{Product group}
}
}
\source{
FAO and other international databases
}
\description{
Harmonized nomenclature for food and agricultural items with codes, names,
and category classifications.
}
\keyword{datasets}
