% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regions_full}
\alias{regions_full}
\title{Regional Classifications}
\format{
A data frame with geographic information:
\describe{
  \item{area_code}{Country/region code}
  \item{area_name}{Country/region name}
  \item{region}{Regional grouping}
  \item{income_group}{World Bank income classification}
}
}
\source{
UN, FAO, World Bank classifications
}
\description{
Country codes and regional aggregations for geographic analysis.
}
\keyword{datasets}
