% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_functions.R
\name{calc_avail_fp_dtm}
\alias{calc_avail_fp_dtm}
\title{Calculate Available Footprint Using Detailed Trade Matrix}
\usage{
calc_avail_fp_dtm(filtered_cbs, df)
}
\arguments{
\item{filtered_cbs}{Filtered CBS data with production and import}

\item{df}{Product footprint data frame}
}
\value{
A data frame with availability footprints including bilateral import footprints
}
\description{
Calculates the footprint of product availability using bilateral trade data (DTM).
}
\examples{
\dontrun{
avail_fp <- calc_avail_fp_dtm(filtered_cbs, product_footprints)
}
}
