% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{drop_cols}
\alias{drop_cols}
\title{Drop columns from data frame}
\usage{
drop_cols(df, ...)
}
\arguments{
\item{df}{Data frame}

\item{...}{Column names to drop}
}
\value{
Data frame without the specified columns
}
\description{
Drops an undefined number of columns from a data frame
}
