% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Modern_variety_adoption}
\alias{Modern_variety_adoption}
\title{Modern Variety Adoption Timeline}
\format{
A data frame with columns:
\describe{
  \item{region_HANPP}{World region name (8 regions).}
  \item{Year}{Year (decadal from 1900 to 2020).}
  \item{Modern_share}{Share of area under modern varieties (0-1).}
  \item{HI_correction_factor}{Multiplicative correction for
    residue:product ratio (>1 = more residue in traditional systems).}
}
}
\source{
Evenson & Gollin (2003) Science 300:758-762.
  Krausmann et al. (2013) Global Env Change 23:1170-1181.
  Pingali (2012) PNAS 109:12302-12308.
}
\description{
Regional time-series of modern (high-yielding) variety adoption shares
and their effect on harvest index. Used to correct residue:product ratios
for historical periods when traditional varieties dominated.
}
\keyword{datasets}
